+++ src/config/settings.ts
@@ -1,0 +1,33 @@
+// Endpoint configuration
+// DEVELOPMENT
+export const REST_API_URL = "http://localhost:8000/event/";
+// STAGING
+// export const REST_API_URL = "https://product-navigator-staging.herokuapp.com/event/"
+// PRODUCTION
+// export const REST_API_URL = "https://product-navigator.herokuapp.com/event/"
+
+
+// User configuration
+export const COOKIE_NAME_USER_ID = 'pnUID';
+
+
+// Task configuration
+export const COOKIE_NAME_TASK_USER_ID = 'pnTaskUID';
+export const COOKIE_NAME_TASK_ID = 'pnTaskID';
+
+export const COOKIE_VALUE_TASK_ID_CONTROL = 'i7sM';
+export const COOKIE_VALUE_TASK_ID_TREATMENT = 'i7tM';
+export const COOKIE_VALUE_MISSING = 'none';
+
+
+// Cookie durability
+export const COOKIE_LIFETIME_1HOUR = 1;
+export const COOKIE_LIFETIME_1DAY = 24;
+export const COOKIE_LIFETIME_1WEEK = 24 * 7;
+export const COOKIE_LIFETIME_1MONTH = 24 * 7 * 4;
+export const COOKIE_LIFETIME_1YEAR = 24 * 7 * 52;
+
+
+// Amazon Constants
+// Platform Constants | Amazon Basics
+export const AMA_URL = 'https://www.amazon';