+++ src/content/amazon/v2/model/Event.js
@@ -1,7 +1,0 @@
-export class Event {
-    constructor(item, type) {
-        this.item = item;
-        this.timestamp = new Date().toJSON();
-        this.type = type;
-    }
-}