+++ src/content/util/cookie.ts
@@ -1,6 +1,7 @@
 // https://www.w3schools.com/js/js_cookies.asp
 
-export function setCookie(cname, cvalue, exhours) {
+export function setCookie(cname: string, cvalue: string, exhours: number) {
+    // exhours = 0 for session cookie
     const d = new Date();
     d.setTime(d.getTime() + (exhours * 60 * 60 * 1000));
     let expires = "expires=" + d.toUTCString();
@@ -8,7 +9,7 @@ export function setCookie(cname, cvalue, exhours) {
     return cvalue;
 }
 
-export function getCookie(cname) {
+export function getCookie(cname: string) {
     let name = cname + "=";
     let ca = document.cookie.split(';');
     for (let i = 0; i < ca.length; i++) {
@@ -21,16 +22,4 @@ export function getCookie(cname) {
         }
     }
     return "";
-}
-
-export function checkCookie() {
-    let user = getCookie("username");
-    if (user != "") {
-        alert("Welcome again " + user);
-    } else {
-        user = prompt("Please enter your name:", "");
-        if (user != "" && user != null) {
-            setCookie("username", user, 365);
-        }
-    }
 }