+++ src/content/util/isInViewport.ts
@@ -1,4 +1,6 @@
-export function isInViewport(htmlElement) {
+// https://stackoverflow.com/a/7557433/11236
+
+export function isInViewport(htmlElement: HTMLElement) {
     const rect = htmlElement.getBoundingClientRect();
     return (
         rect.top >= 0 &&