+++ webpack.config.js
@@ -1,34 +1,28 @@
 const path = require('path');
-const CopyPlugin = require('copy-webpack-plugin');
-const webpack = require('webpack');
+
 
 module.exports = {
     entry: {
-        searchResultsPage: path.resolve(__dirname, "src/content/amazon/v2/searchResultsPage.js"),
-        productDetailPage: path.resolve(__dirname, "src/content/amazon/v2/productDetailPage.js")
+        searchResultsPage: path.resolve(__dirname, "src/content/searchResultsPage.ts"),
+        productDetailPage: path.resolve(__dirname, "src/content/productDetailPage.ts")
     },
     output: {
-        path: path.resolve(__dirname, "dist"),
         filename: "[name].js",
+        path: path.resolve(__dirname, "dist"),
+        clean: true
     },
     resolve: {
-        extensions: [".js"],
+        extensions: ['.tsx', '.ts', '.js'],
     },
     module: {
         rules: [
+            {
+                test: /\.tsx?$/,
+                use: 'ts-loader',
+                exclude: /node_modules/,
+            },
             { test: /\.css$/i, use: ["style-loader", "css-loader"] }
         ],
     },
-    plugins: [
-        new CopyPlugin({
-            patterns: [
-                { from: "public" },
-                { from: path.resolve(__dirname, "src/style"), to: path.resolve(__dirname, "dist/style") }
-            ]
-        }),
-        new webpack.ProvidePlugin({
-            $: "jquery",
-            jQuery: "jquery"
-        })
-    ]
+
 };