+++ webpack.q.js
@@ -7,11 +7,11 @@ const webpack = require('webpack');
 
 module.exports = merge(config, {
     mode: 'development',
-    devtool: 'inline-source-map',
+    devtool: false,
     plugins: [
         new CopyPlugin({
             patterns: [
-                { from: "public/dev" },
+                { from: "public/staging" },
                 { from: path.resolve(__dirname, "src/style"), to: path.resolve(__dirname, "dist/style") }
             ]
         }),